+++ src/Symfony/Component/ObjectMapper/Transform/MapCollection.php
@@ -11,32 +11,36 @@
 
 namespace Symfony\Component\ObjectMapper\Transform;
 
+use Symfony\Component\ObjectMapper\Attribute\Map;
+use Symfony\Component\ObjectMapper\Attribute\MapCollection as MapCollectionAttribute;
 use Symfony\Component\ObjectMapper\Exception\MappingException;
+use Symfony\Component\ObjectMapper\MappingAwareTransformCallableInterface;
 use Symfony\Component\ObjectMapper\ObjectMapper;
 use Symfony\Component\ObjectMapper\ObjectMapperInterface;
-use Symfony\Component\ObjectMapper\TransformCallableInterface;
 
 /**
  * @template T of object
  *
- * @implements TransformCallableInterface<object, T>
+ * @implements MappingAwareTransformCallableInterface<object, T>
  */
-class MapCollection implements TransformCallableInterface
+class MapCollection implements MappingAwareTransformCallableInterface
 {
     public function __construct(
         private ObjectMapperInterface $objectMapper = new ObjectMapper(),
     ) {
     }
 
-    public function __invoke(mixed $value, object $source, ?object $target): mixed
+    public function __invoke(mixed $value, object $source, ?object $target, Map $map): mixed
     {
         if (!is_iterable($value)) {
             throw new MappingException(\sprintf('The MapCollection transform expects an iterable, "%s" given.', get_debug_type($value)));
         }
 
+        $itemClass = $map instanceof MapCollectionAttribute ? $map->itemClass : null;
+
         $values = [];
         foreach ($value as $k => $v) {
-            $values[$k] = $this->objectMapper->map($v);
+            $values[$k] = $this->objectMapper->map($v, $itemClass);
         }
 
         return $values;